+++ trash-app/src/main/java/com/project/trash/auth/config/JwtConfig.java
@@ -6,5 +6,5 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
  * JWT 설정 정보
  */
 @ConfigurationProperties(prefix = "jwt")
-public record JwtConfig(String secretKey, Integer accessExpiration, Integer refreshExpiration) {
+public record JwtConfig(String secretKey, Long accessExpiration, Long refreshExpiration) {
 }
