+++ trash-app/src/main/java/com/project/trash/member/response/AccessTokenInfoResponse.java
@@ -3,9 +3,6 @@ package com.project.trash.member.response;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 
-/**
- * 엑세스 토큰 재발급 반환
- */
 @Getter
 @Schema(title = "엑세스 토큰 재발급 응답")
 public class AccessTokenInfoResponse {
@@ -14,9 +11,9 @@ public class AccessTokenInfoResponse {
   private final String accessToken;
 
   @Schema(description = "엑세스 토큰 유효시간", example = "86400000")
-  private final Integer accessExpiration;
+  private final Long accessExpiration;
 
-  public AccessTokenInfoResponse(String accessToken, Integer accessExpiration) {
+  public AccessTokenInfoResponse(String accessToken, Long accessExpiration) {
     this.accessToken = accessToken;
     this.accessExpiration = accessExpiration;
   }
